Renamed outer CCI effects package
