+++ spateo/tools/CCI_effects_modeling/MuSIC_upstream.py
@@ -1,6 +1,6 @@
 """
 Functionalities to aid in feature selection to characterize signaling patterns from spatial transcriptomics. Given a
-list of signaling molecules (ligands or receptors) and
+list of signaling molecules (ligands or receptors) and/or target genes
 """
 import argparse
 import multiprocessing
