+++ spateo/tools/CCI_effects_modeling/__init__.py
@@ -1,7 +1,7 @@
 """
 General and generalized linear modeling of spatial transcriptomics
 
-Option to call functions from ST_modeling (e.g. st.tl.ST_modeling.Niche_Model) or directly from Spateo (e.g.
+Option to call functions from CCI_effects_modeling (e.g. st.tl.CCI_effects_modeling.Niche_Model) or directly from Spateo (e.g.
 st.tl.Niche_Model).
 """
 from .MuSIC import MuSIC
