+++ spateo/tools/__init__.py
@@ -38,4 +38,4 @@ from .labels import Label, create_label_class
 from .lisa import GM_lag_model, lisa_geo_df, local_moran_i
 from .live_wire import LiveWireSegmentation, compute_shortest_path, live_wire
 from .spatial_degs import cellbin_morani, moran_i
-from .ST_modeling import *
+from .CCI_effects_modeling import *
